+++ examples/miniapps/services/run.py
@@ -22,14 +22,15 @@ if __name__ == '__main__':
     # s3 = boto.s3.connection.S3Connection(aws_access_key_id='KEY',
     #                                      aws_secret_access_key='SECRET')
     #
-    # example.main.main(uid=sys.argv[1],
-    #                   password=sys.argv[2],
-    #                   photo=sys.argv[3],
-    #                   users_service=example.services.Users(logger=logger,
-    #                                                        db=database),
-    #                   auth_service=example.services.Auth(logger=logger,
-    #                                                      db=database,
-    #                                                      token_ttl=3600),
-    #                   photos_service=example.services.Photos(logger=logger,
-    #                                                          db=database,
-    #                                                          s3=s3))
+    # example.main.main(
+    #     uid=sys.argv[1],
+    #     password=sys.argv[2],
+    #     photo=sys.argv[3],
+    #     users_service=example.services.UsersService(logger=logger,
+    #                                                 db=database),
+    #     auth_service=example.services.AuthService(logger=logger,
+    #                                               db=database,
+    #                                               token_ttl=3600),
+    #     photos_service=example.services.PhotosService(logger=logger,
+    #                                                   db=database,
+    #                                                   s3=s3))